refactor: decompose commands/init.py (#561)
Move the project-scaffolding helpers out of the oversized
src/rhiza/commands/init.py into a new
src/rhiza/commands/_init_helpers.py module and re-export them so the
public import surface (rhiza.commands.init._create_*) stays stable.

init.py drops from 320 to 193 statements; _init_helpers.py is 136 —
both under the ~200-statement target. The network/discovery helpers
and init() itself stay in init.py to preserve the test patch targets
at rhiza.commands.init.*; _create_template_file resolves the
name-patchable _get_latest_tag via the init module at call time.

All gates green; test suite 648 passing at 100% coverage.
